Fix: listing volumes with filters
This patch includes 2 fix for the following issues:

1) ast.literal_eval() doesn't work with int and float values
See comment inline

2) Do not traverse and modify the same dict:
while traversing filters dict, we're modifying it inside the loop
which distorts the order of elements (adding modified elements in the
end). to fix this, i've used a temp dict that will be used to traverse
and modification will be done in the filters dict.

Closes-Bug: #1883490

Change-Id: I18b4b0b1b71904b766f7b89df49f5539e3c7662a