Fix: show volume transfer by name for non-admins
When we try to show a volume transfer as a non-admin, it doesn't work.
The problem is that the 'all_tenants' key isn't parsable by the DB,
we remove it for admins but not for non-admins.
Since all_tenants is just used to redirect to the correct db method,
it should be removed for non-admins as well.

Change-Id: I79362ee6b819c6472fa5e69ea1a85a1185dda5a3
Closes-Bug: #1884268
(cherry picked from commit 1f733cdf2dcf014b4391ebd733693d3cbcb1f848)