Disable siblings for the cindelib functional tests
Follow the same pattern of the other jobs which deploy devstack
but runs also functional tests (see devtack-tox-functional
and devstack-tox-functional-consumer).

Without this, the older branches where cinderlib functional tests
use Python 2 (stein, and soon train when that is branched) fails
because they try to install packages part of which are Python 3
only.

The same change was applied to the branchless jobs defined
inside cinder-tempest-plugins: https://review.opendev.org/738078/

Change-Id: I2e8af74bae63c487e786da6e99aa4827029a05a2