Kaminario: Fix unique_fqdn_network option
The Kaminario driver used to accept the unique_fqdn_network
configuration option in the specific driver section, but when we moved
this option to be an option also used by 3PAR we introduced a regression
where the parameter was now only used if it was defined in the shared
configuration group.

This patch fixes this and makes it possible for the option to be defined
in the shared configuration group as well as the driver specific
section.

Closes-Bug: #1886042
Change-Id: I88c5afacff5f1f780c2dc850e3af417b1b081592
(cherry picked from commit 53504f82a53146151935e6aac301a41e2b387623)