PowerMax Driver -  Replacing generations with snap_ids
Historically generations we used for a snapshot series, starting
at 0 and incrementing by 1 for every new generation.  This was
problematic when a snapVx was deleted out of sequence as
each snapVx descendent had to decrement by one.  snap_ids
are unique to each snapVx and will benefit greatly in concurrent
deletions of bootable volumes

Change-Id: I0edf2ac777bef888e760f711a94e3fe4f94262ae
Implements: blueprint powermax-snapset-ids