NEC driver: fix a snapshot detach error
In NEC driver, _fc_terminate_connection() refers to
Volume.volume_attachment of an argument.
However, the argument is a snapshot object when a snapshot is being
detached.
The snapshot object does not have the "volume_attachment" attribute.

This patch avoids referring to the "volume_attachment" attribute when
the argument is a snapshot.

Change-Id: Ibc5fa0901ae9c0766f98d623357bade266598bff
Closes-Bug: #1887885
(cherry picked from commit 6e8dd2f0c01060dbde1a0263da053ffd1a7322f3)