NEC driver: fix live-migration failure with FC
The initialize_connection() in the driver sometimes returns wrong LUN
and causes a live-migration failure.
The function searches the LUN from LD sets (attached hosts) and returns
the first hit.
The function must return an LUN of the destination host, but the first
hit may be an LUN of the source host.

This patch fixes initialize_connection() to return correct LUN by
searching with WWPN of the desitination host.

Change-Id: I102ae84204e0d88814a7d2e028f7cec118ad6b60
Closes-Bug: #1887908
(cherry picked from commit 94c1d241533397bb938682abeaa51ae74d069cb7)