Brocade: Fix AttributeError when raising exception
When we moved Brocade exceptions from cinder.exception to
cinder.zonemanager.drivers.broade.exception we missed one file,
brcd_rest_fc_zone_client, so when we use that file (setting
`fc_southbound_protocol = REST_HTTP`) if it tries to raise an exception
we end up seeing:

  AttributeError: module 'cinder.exception' has no attribute
  'BrocadeZoningRestException'

The CI didn't catch it because there are no unit tests for the
brcd_rest_fc_zone_client.py file.

This patch fixes the issue by changing the import so it gets the
exceptions from the Brocade file and it also adds some tests for the
first half (211 lines) of the code in the brcd_rest_fc_zone_client.py
file.

TrivialFix
Closes-Bug: #1866860
Change-Id: Idfdb8796d5f4b650aa762e54a189c9ab48f93d1a