Add lsscsi to bindep
lsscsi is needed for some of the SCSI device identification. This
identification is actually done in os-brick, but due to the way devstack
and potentially other things install from source, os-brick's bindep.txt
file may not be used.

Since Cinder needs the package in order for os-brick to operate, we can
just add it to cinder so we can make sure it is available.

Change-Id: Iee0a30c4e623c9b22c377093cd600cc8f6da78eb
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>