Fix revert to snapshot for non admins
When a non user calls revert to snapshot using the snapshot name the
client returns a failure status:

  $ cinder --os-volume-api-version 3.59 revert-to-snapshot snap1
    ERROR: No snapshot with a name or ID of 'snap1' exists.

The revert to snapshot API requires an UUID, so the cinderclient lists
the snapshots passing "all_tenants=1" and the name of the snapshot to
find its first when it's not a UUID.

The problem that we were only removing the "all_tenants" filter for
admins, leaving it for normal users and passing it to the DB which
wouldn't understand that filter and return None.

This patch ensure we always remove the "all_tenants" filter (as we do in
other places) before calling the OVO method to list snapshots.

Change-Id: I5fd3d7840b36f6805143cd1b837258232e7bc58a
Fixes-Bug: #1889758
(cherry picked from commit 71a080b22688c2fb15545736e807d233932e22e8)