Change the CLI document for the extend-attached-volume
The extend documentation will need to be updated to indicate that a
volume that is attached to an instance can be extended if the driver has
the capability 'in-use-extend'.

Prior to microversion 3.42 the volume status must be available.
Starting with microversion 3.42, attached volumes with status in-use
may be able to be extended depending on policy and backend storage.

Change-Id: I4d0a439dd53fc7eaee38d0a32fe4e6f0ab4fbf9f
Closes-Bug: #1871759