Remove Train online data migrations
Online data migrations are supposed to be removed in the following
development cycle, so these migrations from Train are overdue.

This change should NOT be backported to Ussuri, however, as there
was a bug in the Train migrations and a fixed version of the migrations
may still need to be run in Ussuri.  See Bug #1893107 for more info.

Change-Id: I78681ea89762790f544178725999903a41d75ad1
Closes-bug: #1893114