Implement revert_to_snapshot() for StorPool.
Use the StorPool API to revert a volume to the specified snapshot
immediately, without deleting and recreating the volume. This also
preserves the volume's metadata, such as e.g. SCSI WWN identifier for
the purposes of recognizing it on VM instances that it has been
previously attached to.

Change-Id: I3b8126cf1706921eac9556add11c558d973cf272
Implements: blueprint storpool-revert-to-snapshot