[api-ref]Add replication_status(optional) to List groups with details response
The response of "list groups with deatils" interface refers to
that of "show group details",whose response contains replication_status.
So it is nesscessary to add replication_status to the response of
"list groups with details", so that user can get replication information
directly instead of calling "show group details" interface again.

https://docs.openstack.org/api-ref/block-storage/v3/index.html?expanded
=list-groups-with-details-detail#generic-volume-groups-groups

Change-Id: I465cab85cd682b44a3b2fba116ea6dbb3f5aa0e9
Closes-Bug: #1894834