Log information about the Ceph v2 clone API
Starting with the Mimic release and the Ceph RBD clone v2 feature,
it is no longer necessary to manage snapshot protection status.
The usage of clone v2 is controlled via the Ceph “rbd default
clone format” configuration option which defaults to “auto”, but
can be overridden to “1” to force the legacy clone v1 behavior or
“2” to force the new clonev2 behavior.

This must be configured in Ceph, it's not something that can be
configured on the Cinder side.

As an alternative, setting the minimum client to mimic
"$ ceph osd set-require-min-compat-client mimic"

Above all with Ceph RBD clone v2 support enabled, image snapshots
can be cloned without marking the snapshot as protected because
RBD automatically tracks the usage of the snapshot by clones. As
a result Ceph RBD clone v2 API improves performance.

This proposed patch adds logging messages for the operator
indicating whether Ceph has been configured to use the Ceph RBD
v2 clone feature.

In this way the operator can consult with the Ceph administrator
to take appropriate action.

You can determine which volumes are using the clone v2 feature
thanks to volume.volume.op_features(). It returns a value depending
on the v2 clone api being enabled or not.

Co-authored-by: Eric Harney <eharney@redhat.com>
Change-Id: Ib52879a270a4ae4cdd3cb5fc18f2b7bdbccd8ab5