NetApp SolidFire: Fix error on cluster workload rebalancing
When SolidFire is under heavy load or being upgraded, the
SolidFire cluster may automatically move connections from primary
to secondary nodes, in order to rebalance cluster workload.

Although this operation ocurrs very quickly, if an operation is made
to a volume at the same time it's being moved, there might be a
chance that API calls such as create snapshot could fail with
xNotPrimary error. Normally this will succeed on a retry of the
operation.

This patch fixes this issue by adding the xNotPrimary exception to
our list of retryable exceptions in the SolidFire driver.

Change-Id: I67dd2bfba37adcb7cda5f1cd08ff641410ec3f6b
Closes-Bug: #1891914