Reno: Use customized template
Even though we have updated the release notes instructions [1] we keep
forgetting to do it, so this patch adds a customized template to
Cinder's release notes to help developers remember.

The template includes a link to the contributor page where we explain
additional details as well as examples for normal and driver features
and fixes.

[1]: https://docs.openstack.org/cinder/latest/contributor/releasenotes.html

Change-Id: I6a57751310d0a52fba43aea3facac9207548f1ef