[SVF]:Fix clone fcmap not being deleted in cleanup
[Spectrum Virtualize Family] check_vdisk_fc_mappings is not deleting
clone fc maps, if the flashcopy status is in copying autodelete is set
to on and copy progress is 100%. when multiple clones and snapshots
are created using common source volume, deleting the source volume
is not removing all the fcmaps involved in mapping with source volume.

This patch fixes the issue by removing the clone fcmap and then
removing the snapshot fcmap when the source has more than one fcmap.

Closes-Bug: #1890254

Change-Id: I84629cde6d27d6889a30d7c13d99f57c660adde7