RBD: Run flatten in a different thread when cloning a volume
The current implementation of create_cloned_volume calls flatten
directly, and this makes whole thread of cinder-volume blocked by that
flatten call. This causes heartbeat timeout in RabbitMQ when cloning
a volume with rbd backend.

This patch makes sure that flatten is executed in a different thread,
to allow heatbeat thread to run while flattening a rbd image.

Closes-Bug: #1898918
Change-Id: I9f28260008117abcebfc96dbe69bf892f5cd14fe