Remove collections.abc backwards compatibility
The collections module moves some abstract classes into the abc
submodules in py3. While we supported older versions of python, we
needed to handle importing from either the old or new locations.

Now that we only support runtimes that include the collections.abc
module, we can remove the backwards compatibility handling we had for
the old location.

Change-Id: Idd106a8199fa586e0b34c054383d64218383c001
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>