Do not fail when depth is greater than rbd_max_clone_depth
Removed a sanity check in the code that raised an exception
if the clone depth of a volume to be cloned exceeded the
rbd_max_clone_depth config value.  A consequence of this
check was that if an operator lowered the value, volumes
whose clone depth was greater than the new value (as would
be allowed by the previous, higher setting) could no longer
be cloned.

Change-Id: I8c445058a25c2eca2fda91bdeb6befedae34ccf2
Closes-bug: #1901241
(cherry picked from commit 8d8d242c3924037e6774a681efabce4893a26646)