Fix unnecessary migration on retype
Retyping a volume with different QoS entries will require a migration,
even if both QoS entries have the exact same data.

This is caused by an incorrect comparison of the QoS information where
we are comparing unnecessary fields: created_at, updated_at, and
deleted_at.

The reason why the unit tests were not picking this up is because the
test was using 'created_at' as one of the spec keys (something artificial
that would not happen in the real world) and it was replacing the value
from the QoS entity.

This patch removes those 3 fields from the QoS comparison, avoiding
unnecessary migrations on retyping.

Close-Bug: #1901188
Change-Id: I979d4db90ee40b4c0a37cd198d7dfcfd7f8d75f7