db: Integrate alembic
Start the move to alembic for real by integrating it into the 'db sync'
command flow. sqlalchemy-migrate is not removed entirely. Not yet
anyway. Instead, we will apply all sqlalchemy-migrate migrations up to
the final '140_create_project_default_volume_type' migration, after
which we will switch over to alembic. In a future release we can remove
the sqlalchemy-migrate migrations and rely entirely on alembic.

Change-Id: Id30929de443d5f7e3923d1061a9f7bbe8b949d5a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>