db: Add initial alembic migration
Re-implement the current base migration from sqlalchemy-migrate in
alembic. This is almost identical, with the main changes being that
everything is handled via alembic-provided APIs (e.g. 'create_table')
rather than APIs that are monkey patched onto sqlalchemy resource
classes by sqlalchemy-migrate (e.g. 'Table.create'). Reviewers are
encouraged to diff the old migration file,
'cinder/db/migrations/versions/921e1a36b076_initial.py', against the new
one, 'cinder/db/legacy_migrations/versions/135_cinder_init.py'.

For now, none of cinder's machinery actually uses this. In a future
change, we will start using this instead of the sqlalchemy-migrate
migrations for new deployments dummy apply this and simply dummy apply
it for existing deployments.

This new migration can be compared to legacy migrations using a similar
technique to that described at [1]. Unlike that example, we don't have
different versions of sqlalchemy-migrate migrations to compare here but
rather two totally different sets of migrations. As such, some tweaks
are needed. Consider the below connection string, which indicates we'll
work with an SQLite database:

  DB_URL=sqlite:///cinder.db

The sqlalchemy-migrate-based migrations can be applied using either the
'migrate' utility or the 'manage.py' script provided in the legacy
migrations directory ('cinder/db/sqlalchemy/legacy_migrations'). We'll
use the former:

  $ cd cinder/db/sqlalchemy/legacy_migrations
  $ migrate version_control --url $DB_URL --repository . --version 134
  $ migrate upgrade --url $DB_URL --repository . 135

(note: we're only applying the initial migration since other migrations
are squashed in in a later patch)

We could presumably write a 'manage.py' equivalent for alembic but we
won't do so here, instead using the 'alembic' utility. This utility
doesn't appear to provide an easy way to override arbitrary config
options from the CLI, which means you must modify the 'alembic.ini'
configuration file provided in the root of the migrations repo before
running 'alembic':

  $ cd cinder/db/sqlalchemy/migrations
  $ crudini --set alembic.ini alembic sqlalchemy.url $DB_URL
  $ alembic upgrade head

With this completed, you can compare the two databases. For this SQLite
example, we use the 'sqldiff' utility from the SQLite contributors [2]
(available as part of 'sqlite-tools' on Fedora):

  $ cd cinder/db/sqlalchemy
  $ sqldiff --schema ../legacy_migrations/cinder.db cinder.db
  CREATE TABLE alembic_version (
          version_num VARCHAR(32) NOT NULL,
          CONSTRAINT alembic_version_pkc PRIMARY KEY (version_num)
  );
  DROP TABLE migrate_version;

From this we can see that we've added a new 'alembic_version' table and
dropped the legacy 'migrate_version'. If we dropped the '--schema' flag
from the command then we'd also see some timestamp differences for the
default entries in some tables, but no more.

[1] https://that.guru/blog/comparing-nova-db-migrations/
[2] https://sqlite.org/sqldiff.html

Change-Id: I3f24313b5260a728dc98d47143ce83dba221c257
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>