RBD: Change rbd_exclusive_cinder_pool's default
In Cinder we always try to have sane defaults, but the current RBD
default for rbd_exclusive_cinder_pools may lead to issues on deployments
with a large number of volumes:

- Cinder taking a long time to start.
- Cinder becoming non-responsive.
- Cinder stats gathering taking longer than the gathering period.

This is cause by the driver making an independent request to get
detailed information on each image to accurately calculate the space
used by the Cinder volumes.

With this patch we change the default to make sure that these issues
don't happen in the most common deployment case (the exclusive Cinder
pool).

Related-Bug: #1704106
Change-Id: I839441a71238cdad540ba8d9d4d18b1f0fa3ee9d