Update secure RBAC check strings with descriptions
This commit attempts to clarify the intended usage of each persona.
After reading the comment for each, it should be clear how and when to
apply them to specific resources. If not, then we should continue to
improve the definition and remove ambiguity.

This commit also adds a new composite check string for system
administrators and project members. This implements a more robust
security posture where system administrators are allowed to do things
that project member can do. Individual deployments can offload some of
the administrative burden to system members if desired by override the
default SYSTEM_ADMIN_OR_PROJECT_MEMBER check string.

Change-Id: I7bad7b5ebaff3d191c0701a9e72ed4c6334ecaa3