PowerMax Driver - Initiator group contents check
Previously, the checking the contents of an existing initiator group
was configurable.  If there was a mismatch between the initiators
from the connector object and the contents of the IG, without this
check, it left the system in an undesirable state.  Now, we check
for the mismatch and throw an exception if one occurs if the
initiator_check is not configured. If initiator_check is True, the
driver recreates the initiator group and masking view and the attach
operation can proceed.

Change-Id: I5e871679817b5aacbdd9aefdc888d1d43202b411