NetApp ONTAP: Fix QoS lost after moving volume
When a Cinder volume is created with a volume-type associated to a QoS
entity, the driver creates a QoS policy group at the ONTAP back end, and
associates it to the entity representing the Cinder volume (either a LUN
or a file within an NFS share).

On NetApp NFS, when a migrate operation is issued and it completes, the
resulting volume ends up without a QoS.  That happens because the file
is being renamed while the QoS refers to the now non-existent file.

This patch makes it so that the file is not renamed when finishing a
migration and the driver code uses the ``name_id`` attribute instead of
the ``id`` one to refer to the right UUID.

Conflicts:
    cinder/tests/unit/volume/drivers/netapp/dataontap/test_nfs_base.py
    cinder/volume/drivers/netapp/dataontap/nfs_base.py
Both conflicts are caused by the new code added by
I507083c3e34e5a5cf1db9a3d1f6bef47bd51a9f8

Closes-Bug: #1906291
Change-Id: Icd7a929e7cbce0c74f6b340f4e09f74a8098d752
(cherry picked from commit da100e1fca610cefb7e77079ab3170b068e328a2)
(cherry picked from commit ee49b67414ef990dd4963f3a4a18d8d3784d8626)