NetApp ONTAP: Fix QoS lost after moving volume
When a Cinder volume is created with a volume-type associated to a QoS
entity, the driver creates a QoS policy group at the ONTAP back end, and
associates it to the entity representing the Cinder volume (either a LUN
or a file within an NFS share).

On NetApp NFS, when a migrate operation is issued and it completes, the
resulting volume ends up without a QoS.  That happens because the file
is being renamed while the QoS refers to the now non-existent file.

This patch makes it so that the file is not renamed when finishing a
migration and the driver code uses the ``name_id`` attribute instead of
the ``id`` one to refer to the right UUID.

Closes-Bug: #1906291
Change-Id: Icd7a929e7cbce0c74f6b340f4e09f74a8098d752