PowerMax Driver - Extend replicated volume
In a replicated scenario it is possible for an R2 volume to be larger
than its paired R1 if we do not force a new volume creation on both
sides of the pair.  While using existing volumes is more efficient it
throws an exception if we want to extend the R1 to a size that is less
than or equal to the R2.

To fix this we are setting the create_new_volumes flag to true for
all replicated volumes.  It will remain false for non replicated volumes.

Change-Id: I667213a1abb025010dc306e4a012dd87c4818756