Hitachi: Wait until the volume can be deleted
In Hitachi Storage, after create cloned volume, it may take time to transition
a volume status in the storage, and it may not be possible to delete the cloned
source volume immediately.
This issue must be resolved for the Hitachi driver to work successfuly
cinder-tempest-plugin.
We add wait processing in the function for checking status to solve this issue.

Change-Id: Ic79066792294d6efb3aa70b5811af4d143d7d271
Closes-Bug: #1908792