Prohibit volume manage to an encrypted volume type
Managing a volume to an encrypted volume type should not be allowed.
One reason is that there is no way for an operator to specify an
encryption key ID for the volume. Another is that we already don't
allow a volume of an encrypted type to be un-managed, so this change
will be symmetric.

Also update and correct the api-ref for this call.

Co-authored-by: xuan <xuanyd@outlook.com>
Co-authored-by: Brian Rosmaita <rosmaita.fossdev@gmail.com>

Change-Id: Ic2da41f3962c1108f974aca952bce3da6d6ac277
Closes-bug: #1944577
(cherry picked from commit 8088dc9580356ca51581c56888c9022848f543b1)
change:
  cinder/tests/unit/api/contrib/test_volume_manage.py - used a
  feature of unittest.mock that was introduced in py38; rewritten to
  be compatible with py36
(cherry picked from commit 9c1aa92b9578c1dc1d63d4d9124d873438d495eb)