[stable-only] Cap bandit and fix lower-constrains
bandit 1.6.3 [1] release has dropped support for py2 [2] but the
release is faulty and pip still picks it up for py2 [3][4], so cap to
1.6.2 when using py2.

With the new pip dependency resolver (introduced in pip 20.3) the
lower-constraints job started to timeout and fail. This patch fixes
package version constraint contradictions and add some new lower
constraints to speed up pip's dependency resolution.

[1] https://github.com/PyCQA/bandit/releases/tag/1.6.3
[2] https://github.com/PyCQA/bandit/pull/615
[3] https://github.com/PyCQA/bandit/issues/663
[4] https://github.com/PyCQA/bandit/issues/665

Change-Id: Ib3ff3bd2df8954407e880a759082ef1eee3e76a1