Pure: Add default value to pure_host_personality
Pure's "pure_host_personality" configuration option currently defaults
to a non valid value according to its choices.

This affects tools leveraging the "get_driver_options" mechanism, since
the default value for the option was not considered a valid option.

This patch sets the default explicitly and adds it as a valid choice.

Change-Id: I381b80920d812a3f3b5f75b6dae66aa932f6baa4