PowerMax Driver - Allow for volume uuid in manage volume
In PowerMax, we require a 5 digit hex device id as input for an existing
volume to be managed into OpenStack. However, tempest will not be aware
of this driver specific device id and relies on the uuid of the volume
instead.  Adding logic so that tempest will be able to perform manage/
unmanage suite of tests.

Closes-Bug: #1912751
Change-Id: If94d460b6cec0bf3d102cd687d74474b46bd176b