Cinder - Creating clone of encrypted volume fails
After a clone of an encrypted volume is created, an attach is attempted.
PowerMax driver requires the provider_location be populated in order
to find the volume to attach. For this, the volume object needs to be
updated.

Closes-Bug: #1913054
Change-Id: Idf5b3783ddc333d6d60f28a3d08e5fd28e5c1fa8
(cherry picked from commit e1f56a7c85345883fe030a9298fd76b1dcf63cbf)