Reset state robustification for volume os-reset_status
This commit changes the os-reset_status volume-api
to handle the below use cases
1.Can not reset volume to "ing" states

2.available volume reset to in-use - 400

The _reset_status method is override in VolumeAdminController class
to handle different volume states.

NOTE: with this change, some requests that previously succeeded will
return a 400.  However, these were requests that would leave a resource
in an inconsistent state with an inaccurate status, so a 400 is
appropriate.  We use a 400 (not 409) to be consistent with the current
Block Storage API. This change is not microversioned because it doesn't
make sense to only prevent operators from shooting themselves in the
foot if they happen to specify an appropriate microversion,
and be grievously wounded otherwise.

Partial implement blueprint reset-state-robustification

API state reference sheet
https://wiki.openstack.org/wiki/Reset_State_Robustification

Change-Id: I0a53dfee61dc6372dce53445d4b252304c83c58a